refactor(api): separate HTMX route responsibilities
Move verification workflow coordination and HTMX response rendering out of the route module so handlers focus on HTTP input and response selection.
